refactor(root): Fix compilation issue
