+++ Engine/Shared/SceneManager/ASceneManager.cpp
@@ -151,17 +151,13 @@ void SceneManager::ASceneManager::_initialiseDefaultComponents()
     _registry->register_component<IComponent>(ColourComponent().getType());
     _registry->register_component<IComponent>(FontPathComponent().getType());
     _registry->register_component<IComponent>(MaterialMapComponent().getType());
-    _registry->register_component<IComponent>(MusicPathComponent().getType());
-    _registry->register_component<IComponent>(MusicPitchComponent().getType());
-    _registry->register_component<IComponent>(MusicVolumeComponent().getType());
+    _registry->register_component<IComponent>(MusicComponent().getType());
     _registry->register_component<IComponent>(ObjPathComponent().getType());
     _registry->register_component<IComponent>(Position2DComponent().getType());
     _registry->register_component<IComponent>(Position3DComponent().getType());
     _registry->register_component<IComponent>(ScaleComponent().getType());
     _registry->register_component<IComponent>(Size1DComponent().getType());
-    _registry->register_component<IComponent>(SoundPathComponent().getType());
-    _registry->register_component<IComponent>(SoundPitchComponent().getType());
-    _registry->register_component<IComponent>(SoundVolumeComponent().getType());
+    _registry->register_component<IComponent>(SoundComponent().getType());
     _registry->register_component<IComponent>(TextComponent().getType());
     _registry->register_component<IComponent>(TexturePathComponent().getType());
     _registry->register_component<IComponent>(TextureRectComponent().getType());
