+++ Engine/Shared/SceneManager/ASceneManager.hpp
@@ -24,15 +24,11 @@
 #include "FontPath/FontPathComponent.hpp"
 #include "Editable/EditableComponent.hpp"
 #include "Checkable/CheckableComponent.hpp"
-#include "MusicPath/MusicPathComponent.hpp"
-#include "SoundPath/SoundPathComponent.hpp"
-#include "MusicPitch/MusicPitchComponent.hpp"
+#include "Music/MusicComponent.hpp"
+#include "Sound/SoundComponent.hpp"
 #include "Position2D/Position2DComponent.hpp"
 #include "Position3D/Position3DComponent.hpp"
-#include "SoundPitch/SoundPitchComponent.hpp"
 #include "MaterialMap/MaterialMapComponent.hpp"
-#include "MusicVolume/MusicComponent.hpp"
-#include "SoundVolume/SoundComponent.hpp"
 #include "TexturePath/TexturePathComponent.hpp"
 #include "TextureRect/TextureRectComponent.hpp"
 #include "../DefaultEventHandlers/IEventHandler.hpp"
